Move HubSenderConfiguration to infrastructure
