--- source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Application/Configuration/IHubSenderConfiguration.cs
@@ -1,31 +1,0 @@
-﻿// Copyright 2020 Energinet DataHub A/S
-//
-// Licensed under the Apache License, Version 2.0 (the "License2");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using GreenEnergyHub.Charges.Domain.MarketParticipants;
-
-namespace GreenEnergyHub.Charges.Application
-{
-    /// <summary>
-    /// Interface used to determine which ID and code should be used when sending
-    /// message from the hub out onto the world
-    /// </summary>
-    public interface IHubSenderConfiguration
-    {
-        /// <summary>
-        /// Retrieves the market participant used to send messages from the hub
-        /// </summary>
-        /// <returns>The market participant to use</returns>
-        MarketParticipant GetSenderMarketParticipant();
-    }
-}
